+++ asf_heat_pump_suitability/getters/load_geodata.py
@@ -1,4 +1,5 @@
 import polars as pl
+import geopandas as gpd
 import regex as re
 import os
 
@@ -27,3 +28,20 @@ def load_df_osopen_uprn(**kwargs) -> pl.DataFrame:
     )
 
     return df
+
+
+def load_gdf_plymouth_heat_network_zone_areas(**kwargs) -> gpd.GeoDataFrame:
+    """
+    Load GeoDataFrame with Plymouth heat network zone polygons from Plymouth City Council (CRS: EPSG:27700).
+
+    Args:
+        **kwargs for `gpd.read_file()`
+
+    Returns:
+        gpd.GeoDataFrame: polygons of heat network zones in Plymouth.
+    """
+    gdf = base_getters.get_gdf_from_gpkg_s3_path(
+        path=config["data"]["geodata"]["heat_network_zones"]["plymouth"],
+        **kwargs,
+    )
+    return gdf
